day02/requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
openai==1.6.1
streamlit==1.29.0
plotly==5.18.0
